+++ astropy/units/quantity.py
@@ -21,21 +21,22 @@ import numpy as np
 from ..extern import six
 from ..extern.six.moves import zip
 from .core import (Unit, dimensionless_unscaled, get_current_unit_registry,
-                   UnitBase, UnitsError, UnitConversionError, UnitTypeError)
+                   UnitBase, UnitsError, UnitTypeError)
 from .format.latex import Latex
+from ..utils.compat import NUMPY_LT_1_13
 from ..utils.compat.misc import override__dir__
 from ..utils.compat.numpy import matmul
 from ..utils.misc import isiterable, InheritDocstrings
 from ..utils.data_info import ParentDtypeInfo
 from .. import config as _config
-
+from .quantity_helper import (converters_and_unit, can_have_arbitrary_unit,
+                              check_output)
 
 __all__ = ["Quantity", "SpecificTypeQuantity", "QuantityInfo"]
 
 
 # We don't want to run doctests in the docstrings we inherit from Numpy
 __doctest_skip__ = ['Quantity.*']
-
 
 _UNIT_NOT_INITIALISED = "(Unit not initialised)"
 _UFUNCS_FILTER_WARNINGS = {np.arcsin, np.arccos, np.arccosh, np.arctanh}
@@ -51,23 +52,6 @@ class Conf(_config.ConfigNamespace):
         'negative number means that the value will instead be whatever numpy '
         'gets from get_printoptions.')
 conf = Conf()
-
-
-def _can_have_arbitrary_unit(value):
-    """Test whether the items in value can have arbitrary units
-
-    Numbers whose value does not change upon a unit change, i.e.,
-    zero, infinity, or not-a-number
-
-    Parameters
-    ----------
-    value : number or array
-
-    Returns
-    -------
-    `True` if each member is either zero or not finite, `False` otherwise
-    """
-    return np.all(np.logical_or(np.equal(value, 0.), ~np.isfinite(value)))
 
 
 class QuantityIterator(object):
@@ -416,123 +400,40 @@ class Quantity(np.ndarray):
         # attributes to, etc. After this is called, then the ufunc is called
         # and the values in this empty array are set.
 
+        # In principle, this should not be needed any more in numpy >= 1.13,
+        # but it is still called in some np.linalg modules.
+
         # If no context is set, just return the input
         if context is None:
             return obj
 
         # Find out which ufunc is being used
         function = context[0]
-
-        from .quantity_helper import UNSUPPORTED_UFUNCS, UFUNC_HELPERS
-
-        # Check whether we even support this ufunc
-        if function in UNSUPPORTED_UFUNCS:
-            raise TypeError("Cannot use function '{0}' with quantities"
-                            .format(function.__name__))
-
-        # Now find out what arguments were passed to the ufunc, usually, this
-        # will include at least the present object, and another, which could
-        # be a Quantity, or a Numpy array, etc. when using two-argument ufuncs.
         args = context[1][:function.nin]
-        units = [getattr(arg, 'unit', None) for arg in args]
-
-        # If the ufunc is supported, then we call a helper function (defined
-        # in quantity_helper.py) which returns the scale by which the inputs
-        # should be multiplied before being passed to the ufunc, as well as
-        # the unit the output from the ufunc will have.
-        if function in UFUNC_HELPERS:
-            converters, result_unit = UFUNC_HELPERS[function](function, *units)
-            if function.nout > 1:
-                result_unit = result_unit[context[2]]
-        else:
-            raise TypeError("Unknown ufunc {0}.  Please raise issue on "
-                            "https://github.com/astropy/astropy"
-                            .format(function.__name__))
-
-        if any(converter is False for converter in converters):
-            # for two-argument ufuncs with a quantity and a non-quantity,
-            # the quantity normally needs to be dimensionless, *except*
-            # if the non-quantity can have arbitrary unit, i.e., when it
-            # is all zero, infinity or NaN.  In that case, the non-quantity
-            # can just have the unit of the quantity
-            # (this allows, e.g., `q > 0.` independent of unit)
-            maybe_arbitrary_arg = args[converters.index(False)]
-            try:
-                if _can_have_arbitrary_unit(maybe_arbitrary_arg):
-                    converters = [None, None]
-                else:
-                    raise UnitsError("Can only apply '{0}' function to "
-                                     "dimensionless quantities when other "
-                                     "argument is not a quantity (unless the "
-                                     "latter is all zero/infinity/nan)"
-                                     .format(function.__name__))
-            except TypeError:
-                # _can_have_arbitrary_unit failed: arg could not be compared
-                # with zero or checked to be finite.  Then, ufunc will fail too.
-                raise TypeError("Unsupported operand type(s) for ufunc {0}: "
-                                "'{1}' and '{2}'"
-                                .format(function.__name__,
-                                        args[0].__class__.__name__,
-                                        args[1].__class__.__name__))
-
-        # In the case of np.power and np.float_power, the unit itself needs to
-        # be modified by an amount that depends on one of the input values,
-        # so we need to treat this as a special case.
-        # TODO: find a better way to deal with this.
-        if result_unit is False:
-            if units[0] is None or units[0] == dimensionless_unscaled:
-                result_unit = dimensionless_unscaled
-            else:
-                if units[1] is None:
-                    p = args[1]
-                else:
-                    p = args[1].to_value(dimensionless_unscaled)
-
-                try:
-                    result_unit = units[0] ** p
-                except ValueError as exc:
-                    # Changing the unit does not work for, e.g., array-shaped
-                    # power, but this is OK if we're (scaled) dimensionless.
-                    try:
-                        converters[0] = units[0]._get_converter(
-                            dimensionless_unscaled)
-                    except UnitConversionError:
-                        raise exc
-                    else:
-                        result_unit = dimensionless_unscaled
+        # determine required converter functions -- to bring the unit of the
+        # input to that expected (e.g., radian for np.sin), or to get
+        # consistent units between two inputs (e.g., in np.add) --
+        # and the unit of the result
+        converters, result_unit = converters_and_unit(function, '__call__',
+                                                      *args)
+        if function.nout > 1:
+            result_unit = result_unit[context[2]]
 
         # We now prepare the output object
         if self is obj:
-
             # this happens if the output object is self, which happens
             # for in-place operations such as q1 += q2
 
-            # In some cases, the result of a ufunc should be a plain Numpy
-            # array, which we can't do if we are doing an in-place operation.
-            if result_unit is None:
-                raise TypeError("Cannot store non-quantity output from {0} "
-                                "function in {1} instance"
-                                .format(function.__name__, type(self)))
-
-            if self.__quantity_subclass__(result_unit)[0] is not type(self):
-                raise UnitTypeError(
-                    "Cannot store output with unit '{0}' from {1} function "
-                    "in {2} instance.  Use {3} instance instead."
-                    .format(result_unit, function.__name__, type(self),
-                            self.__quantity_subclass__(result_unit)[0]))
-
-            # If the Quantity has an integer dtype, in-place operations are
-            # dangerous because in some cases the quantity will be e.g.
-            # decomposed, which involves being scaled by a float, but since
-            # the array is an integer the output then gets converted to an int
-            # and truncated.
-            result_dtype = np.result_type(*((args + (float,))
-                                            if any(converters) else args))
-            if not np.can_cast(result_dtype, obj.dtype, casting='same_kind'):
-                raise TypeError("Arguments cannot be cast safely to inplace "
-                                "output with dtype={0}".format(self.dtype))
-
-            result = self  # no view needed since we return the object itself
+            # Check that we're not trying to store a plain Numpy array or a
+            # Quantity with an inconsistent unit (e.g., not angular for Angle),
+            # and that we can handle the type (e.g., that we are not int when
+            # float is required).
+            check_output(obj, result_unit, (args + tuple(
+                (np.float16 if converter and converter(1.) % 1. != 0.
+                 else np.int8)
+                for converter in converters)),
+                         function=function)
+            result = self  # no view needed since already a Quantity.
 
             # in principle, if self is also an argument, it could be rescaled
             # here, since it won't be needed anymore.  But maybe not change
@@ -542,8 +443,8 @@ class Quantity(np.ndarray):
 
             result = self._new_view(obj, result_unit)
 
-        # We now need to treat the case where the inputs have to be scaled -
-        # the issue is that we can't actually scale the inputs since that
+        # We now need to treat the case where the inputs have to be converted -
+        # the issue is that we can't actually convert the inputs since that
         # would be changing the objects passed to the ufunc, which would not
         # be expected by the user.
         if any(converters):
@@ -572,7 +473,7 @@ class Quantity(np.ndarray):
                 else:
                     result._contiguous = self.copy()
 
-            # ensure we remember the scales we need
+            # ensure we remember the converter functions we need
             result._converters = converters
 
             if function in _UFUNCS_FILTER_WARNINGS:
@@ -678,10 +579,99 @@ class Quantity(np.ndarray):
             else:
                 return obj
 
-    def __deepcopy__(self, memo):
-        # If we don't define this, ``copy.deepcopy(quantity)`` will
-        # return a bare Numpy array.
-        return self.copy()
+    def __array_ufunc__(self, function, method, *inputs, **kwargs):
+        """Wrap numpy ufuncs, taking care of units.
+
+        Parameters
+        ----------
+        function : callable
+            ufunc to wrap.
+        method : str
+            Ufunc method: ``__call__``, ``at``, ``reduce``, etc.
+        inputs : tuple
+            Input arrays.
+        kwargs : keyword arguments
+            As passed on, with ``out`` containing possible quantity output.
+
+        Returns
+        -------
+        result : `~astropy.units.Quantity`
+            Results of the ufunc, with the unit set properly.
+        """
+        # Determine required conversion functions -- to bring the unit of the
+        # input to that expected (e.g., radian for np.sin), or to get
+        # consistent units between two inputs (e.g., in np.add) --
+        # and the unit of the result (or tuple of units for nout > 1).
+        converters, unit = converters_and_unit(function, method, *inputs)
+
+        out = kwargs.get('out', None)
+        # Avoid loop back by turning any Quantity output into array views.
+        if out is not None:
+            # If pre-allocated output is used, check it is suitable.
+            # This also returns array view, to ensure we don't loop back.
+            if function.nout == 1:
+                out = out[0]
+            out_array = check_output(out, unit, inputs, function=function)
+            # Ensure output argument remains a tuple.
+            kwargs['out'] = (out_array,) if function.nout == 1 else out_array
+
+        # Same for inputs, but here also convert if necessary.
+        arrays = tuple((converter(input_.value) if converter else
+                        getattr(input_, 'value', input_))
+                       for input_, converter in zip(inputs, converters))
+
+        # Call our superclass's __array_ufunc__
+        result = super(Quantity, self).__array_ufunc__(function, method,
+                                                       *arrays, **kwargs)
+        # If unit is None, a plain array is expected (e.g., comparisons), which
+        # means we're done.
+        # We're also done if the result was None (for method 'at') or
+        # NotImplemented, which can happen if other inputs/outputs override
+        # __array_ufunc__; hopefully, they can then deal with us.
+        if unit is None or result is None or result is NotImplemented:
+            return result
+
+        return self._result_as_quantity(result, unit, out)
+
+    def _result_as_quantity(self, result, unit, out):
+        """Turn result into a quantity with the given unit.
+
+        If no output is given, it will take a view of the array as a quantity,
+        and set the unit.  If output is given, those should be quantity views
+        of the result arrays, and the function will just set the unit.
+
+        Parameters
+        ----------
+        result : `~numpy.ndarray` or tuple of `~numpy.ndarray`
+            Array(s) which need to be turned into quantity.
+        unit : `~astropy.units.Unit` or None
+            Unit for the quantities to be returned (or `None` if the result
+            should not be a quantity).  Should be tuple if result is a tuple.
+        out : `~astropy.units.Quantity` or None
+            Possible output quantity. Should be `None` or a tuple if result
+            is a tuple.
+
+        Returns
+        -------
+        out : `~astropy.units.Quantity`
+           With units set.
+        """
+        if isinstance(result, tuple):
+            if out is None:
+                out = (None,) * len(result)
+            return tuple(self._result_as_quantity(result_, unit_, out_)
+                         for (result_, unit_, out_) in
+                         zip(result, unit, out))
+
+        if out is None:
+            # View the result array as a Quantity with the proper unit.
+            return result if unit is None else self._new_view(result, unit)
+
+        # For given output, just set the unit. We know the unit is not None and
+        # the output is of the correct Quantity subclass, as it was passed
+        # through check_output.
+        out._set_unit(unit)
+        return out
 
     def __quantity_subclass__(self, unit):
         """
@@ -778,6 +768,11 @@ class Quantity(np.ndarray):
                     .format(type(self).__name__, UnitBase, type(unit)))
 
         self._unit = unit
+
+    def __deepcopy__(self, memo):
+        # If we don't define this, ``copy.deepcopy(quantity)`` will
+        # return a bare Numpy array.
+        return self.copy()
 
     def __reduce__(self):
         # patch to pickle Quantity objects (ndarray subclasses), see
@@ -1104,12 +1099,21 @@ class Quantity(np.ndarray):
         result_array = matmul(getattr(other, 'value', other), self.value)
         return self._new_view(result_array, result_unit)
 
-    def __pos__(self):
-        """
-        Plus the quantity. This is implemented in case users use +q where q is
-        a quantity.  (Required for scalar case.)
-        """
-        return self.copy()
+    if NUMPY_LT_1_13:
+        # Pre-numpy 1.13, there was no np.positive ufunc and the copy done
+        # by ndarray did not properly work for scalar quantities.
+        def __pos__(self):
+            """Plus the quantity."""
+            return self.copy()
+
+    else:
+        # In numpy 1.13, a np.positive ufunc exists, but ndarray.__pos__
+        # does not yet go through it, so we still need to define it, to allow
+        # subclasses to override it inside __array_ufunc__.
+        # Presumably, this can eventually be removed.
+        def __pos__(self):
+            """Plus the quantity."""
+            return np.positive(self)
 
     # other overrides of special functions
     def __hash__(self):
@@ -1372,7 +1376,7 @@ class Quantity(np.ndarray):
             try:
                 _value = dimensionless_unscaled.to(self.unit, value)
             except UnitsError as exc:
-                if _can_have_arbitrary_unit(value):
+                if can_have_arbitrary_unit(value):
                     _value = value
                 else:
                     raise exc
@@ -1466,60 +1470,50 @@ class Quantity(np.ndarray):
     # We use the corresponding numpy functions to evaluate the results, since
     # the methods do not always allow calling with keyword arguments.
     # For instance, np.array([0.,2.]).clip(a_min=0., a_max=1.) gives
-    # TypeError: 'a_max' is an invalid keyword argument for this function
+    # TypeError: 'a_max' is an invalid keyword argument for this function.
     def _wrap_function(self, function, *args, **kwargs):
-        """Wrap a numpy function, returning a Quantity with the proper unit
+        """Wrap a numpy function that processes self, returning a Quantity.
 
         Parameters
         ----------
         function : callable
-            numpy function to wrap
+            Numpy function to wrap.
         args : positional arguments
-            any positional arguments to the function.
+            Any positional arguments to the function beyond the first argument
+            (which will be set to ``self``).
         kwargs : keyword arguments
             Keyword arguments to the function.
 
         If present, the following arguments are treated specially:
 
-        unit : `~astropy.units.Unit` or `None`
-            unit of the output result.  If not given or `None` (default),
-            the unit of `self`.
+        unit : `~astropy.units.Unit`
+            Unit of the output result.  If not given, the unit of ``self``.
         out : `~astropy.units.Quantity`
             A Quantity instance in which to store the output.
 
         Notes
         -----
-        Output should always be assigned via a keyword argument.
+        Output should always be assigned via a keyword argument, otherwise
+        no proper account of the unit is taken.
 
         Returns
         -------
         out : `~astropy.units.Quantity`
             Result of the function call, with the unit set properly.
         """
-
-        unit = kwargs.pop('unit', None)
-        out = kwargs.get('out', None)
+        unit = kwargs.pop('unit', self.unit)
+        out = kwargs.pop('out', None)
+        # Ensure we don't loop back by turning any Quantity into array views.
+        args = (self.value,) + tuple((arg.value if isinstance(arg, Quantity)
+                                      else arg) for arg in args)
         if out is not None:
-            if unit is None:
-                unit = self.unit
-
-            if (isinstance(out, Quantity) and
-                out.__quantity_subclass__(unit)[0] is type(out)):
-                # Set out to ndarray view to prevent calling __array_prepare__.
-                kwargs['out'] = out.view(np.ndarray)
-            else:
-                ok_class =  (out.__quantity_subclass__(out, unit)[0]
-                             if isinstance(out, Quantity) else Quantity)
-                raise UnitTypeError("out cannot be assigned to a {0} instance; "
-                                    "use a {1} instance instead.".format(
-                                        out.__class__, ok_class))
-
-        value = function(self.view(np.ndarray), *args, **kwargs)
-        if out is None:
-            return self._new_view(value, unit)
-        else:
-            out._set_unit(unit)
-            return out
+            # If pre-allocated output is used, check it is suitable.
+            # This also returns array view, to ensure we don't loop back.
+            arrays = tuple(arg for arg in args if isinstance(arg, np.ndarray))
+            kwargs['out'] = check_output(out, unit, arrays, function=function)
+        # Apply the function and turn it back into a Quantity.
+        result = function(*args, **kwargs)
+        return self._result_as_quantity(result, unit, out)
 
     def clip(self, a_min, a_max, out=None):
         return self._wrap_function(np.clip, self._to_own_unit(a_min),
@@ -1585,7 +1579,7 @@ class Quantity(np.ndarray):
         raise NotImplementedError("cannot evaluate truth value of quantities. "
                                   "Evaluate array with q.value.any(...)")
 
-    # Calculation --numpy functions that can be overridden with methods
+    # Calculation: numpy functions that can be overridden with methods.
 
     def diff(self, n=1, axis=-1):
         return self._wrap_function(np.diff, n, axis)